+++ golden-ami-lookup/outputs.tf
@@ -1,0 +1,4 @@
+// The ARN of the AMI
+output "ami_id" {
+  value = data.aws_ami.golden_ami.image_id
+}