+++ newrelic/ec2-agent-cloud-config/outputs.tf
@@ -1,0 +1,3 @@
+output "cloudconfig_rendered" {
+  value = data.template_file.cloudconfig.rendered
+}